Moving theming into tokens config
